requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0